requirements.txt
flask
jellyfin-apiclient-python
python-dotenv
python-dateutil
requests
